add pvp toggle and other enhancements
https://www.curseforge.com/wow/addons/autocrop?comment=45

all credit to original authors and commentor/developer zenadar

I've a modified version of this addon with the following changes:

 

1. Changing items is now event based (instead of a time interval).
2. Gas cloud helm is now enabled by providing a valid item ID number. This is to resolve issues that will arise when the other engineering helms are added to the  game.
3. Gas could helms will equip when entering a zone with gas clouds and unequip when entering a zone without them.

 

I made the change in (1) to resolve an issue I was having on my warrior. Specifically, when charging off my mount if the time between when I charge or dismount and when I enter combat is less than the interval check, then I'd be in combat with my riding gear on. You could of course decrease the time interval, but an event based method seems like it might be better solution.

 

See here for the source: https://www.codepile.net/pile/kep3G0nz

 

Chromie, do you have a git-repo for this? If so, can you make a branch for these changes to be pushed to?

Edit: For those asking how to update the addon, do the following:

1. Go to the AutoCrop folder (for my Windows 10 machine it's: C:\Program Files (x86)\World of Warcraft\_classic_\Interface\AddOns\AutoCrop)
2.Open autocrop.lua in a text editor (notepad should work) and replace everything in with the code in the following link (https://www.codepile.net/pile/kep3G0nz) and save the file.
3. Launch wow and when you're in-game do: /ac reset
4. You should get a message about the database being reset. Update your settings (i.e. gas helm).

 

Edit 8/15/2021:

Code in the CodePile link has been updated. Includes functionality for:

- Automatically disabling swapping in PvP (BGs and arena)

- The ability to define zone names to automatically disable the addon (e.g. if you don't want the addon to swap your gear in ZA)

- If you enter combat while mounted and then dismount (meaning you're in combat in your riding set) it will now automatically equip your normal set once you leave combat.